Catch IO MalformedInputExceptions and throw them again with added info
Move custom exceptions to a separate folder instead of scattering them around the code.
